Whoops, didn't mean to delete the icon :D
btw, somehow this passed my own tests and my beta tests, but there is a
typo, in case you want to clone & fix it for yourself: the second occurrence
of the word "villain" is missing an L (must not have pressed hard enough
lol).  But, I'm moving on to the next game, so if you want to change it,
feel free to clone the repo and do that.